refactor: use default type-faces
You know what?  I give up.
Fonts are *huge*!
I'm already looking at my Git diff in the buffer above and seeing an
average of 330KiB extra bytes for Fira GO and 60KiB for JetBrain Mono.

Even on average American-grade Internet connectivity, it takes a while
for the fonts to be pulled and applied.
Looks easy when serving locally to write but it's not helping otherwise.

It's time I gave the user/reader a little more freedom.