quip: culling
No one cares or knows who D.J. Cheese-Like Product is because I made
him up while snipping open a packet of "cheese" for the shells.

This site is over HTTPS so it might not even be accessible over
Netscape 1.0.

No, we do not need more lasers despite the what the retired War
Minister keeps saying.
That's out of budget, we're facing an iron shortage, and the biters
are taking up that patch of iron over there.

Yes, Painkiller is a rad album but this is not really a sensical place
for it after the charm of being in the quips wore off because the RNG
class kept picking it time after time.

Dreams no longer yell at you.
That was made illegal after the sliced bread was relegalized in 1947.
Probably would be best to not go ham on the caps lock key.
Doesn't help the key that actually says "control" is caps lock now and
gets smacked by my palm.

You won't understand it when you're younger either because it still
wouldn't make sense when you're older.