style: tables
I had the awful class in college back in the day and by "back in the
day" I mean early 2019.
We had this web dev class that still heavily insisted tables were the
standard way of structuring a site and because lectures were
monkey-see-monkey-do where you had to repeat the teacher and turn THAT
in and because he leaned so heavily in to that, you tried to do better
and you get left in the dust.

To make things worse in the PHP part (as if it couldn't get worse than
PHP) is that MD5 is "still a good password hashing algorithm". 🤣