chore(git): ignore _drafts/
Ultimately the goal is to be able to commit without having so many
things unstaged and modified.
This becomes a bit of an informational clutter and may not reflect
what is actually published to my site.
Currently the Git status has a way to go before it's clean but this
would help.