quip: I am not the lead guitarist of Tenacious D.
This comparison's plagued me since high school.
Back in the digital media class, they slapped my face on the Nacho
Libre moster back when that movie was a thing.
Even the shills for whatever companies that come knocking door-to-door
here will blurt that out.

An overweight white guy does not make a Jack Black.